Item pickups w/ magnet + sounds!!
- created scripts to allow for easy item creation that will be able to be picked up and attracted by player magnet collider

- setup audio manager script and game object

- lots of new scripts to allow items to work properly

- created a temporary main menu

- added a temporary player sprite

- created two test items fully working with audio (coin and red gem)

Notes: lots of progress, took a couple days or so to figure out how to get items working with being picked up and attracted to the player. watched plenty of tutorials on youtube trying to find the best way and ended up settling on using events which ive never even heard of before. found a good program to use to create sound effects and music :D i look forward to working on the game more!! not sure what ill move to next, thinking maybe starting on the inventory but we will see